+++ Sources/CustomDump/Conformances/Foundation.swift
@@ -325,9 +325,3 @@ extension UUID: CustomDumpStringConvertible {
     "UUID(\(self.uuidString))"
   }
 }
-
-extension AnyHashable: CustomDumpRepresentable {
-  public var customDumpValue: Any {
-    base
-  }
-}
