+++ Tests/CustomDumpTests/Conformances/FoundationTests.swift
@@ -661,22 +661,4 @@ final class FoundationTests: XCTestCase {
       """
     )
   }
-
-  func testAnyHashable() {
-    let user: AnyHashable = HashableUser(id: 1, name: "James")
-    var dump = ""
-    customDump(
-      user,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      HashableUser(
-        id: 1,
-        name: "James"
-      )
-      """
-    )
-  }
 }
